+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/adapter-interfaces
 
+## 0.11.4
+
+### Patch Changes
+
+- Add samlp specific validation for application
+
 ## 0.11.3
 
 ### Patch Changes